+++ examples/actix_web_example/src/main.rs
@@ -69,19 +69,17 @@ async fn test() -> Result<HttpResponse, Error> {
         1u32, 11, 9, 20, 3, 5, 2, 6, 19, 8, 9, 10, 12, 13, 14, 15, 16, 17, 18, 4,
     ];
 
-    let pool_ref = POOL
+    let conn = POOL
         .get()
         .await
         .map_err(|_| ErrorInternalServerError("lol"))?;
 
-    let (client, _statements) = &*pool_ref;
-
-    let st = client
+    let st = conn
         .prepare_typed("SELECT * FROM topics WHERE id=ANY($1)", &[Type::OID_ARRAY])
         .await
         .expect("Failed to prepare");
 
-    let (t, _u) = client
+    let (t, _u) = conn
         .query_raw(
             &st,
             [&ids as &(dyn ToSql + Sync)]
@@ -114,8 +112,6 @@ async fn test() -> Result<HttpResponse, Error> {
         .await
         .map_err(|_| ErrorInternalServerError("lol"))?;
 
-    drop(pool_ref);
-
     Ok(HttpResponse::Ok().json(&t))
 }
 
@@ -124,7 +120,6 @@ async fn test_redis(pool: web::Data<Pool<RedisManager>>) -> Result<HttpResponse,
         .get()
         .await
         .map_err(|_| ErrorInternalServerError("lol"))?
-        .get_conn()
         .clone();
 
     redis::cmd("PING")