+++ examples/rocket_example/src/main.rs
@@ -78,17 +78,14 @@ async fn index(
     pool: State<'_, Pool<PostgresManager<tokio_postgres::NoTls>>>,
 ) -> Result<content::Json<String>, Debug<std::io::Error>> {
     // pool.get return the Conn and a reference of pool so that we can use the connection outside a closure.
-    let pool_ref = pool.get().await.map_err(MyError::from)?;
+    let conn = pool.get().await.map_err(MyError::from)?;
 
-    // deref or deref mut to get connection from pool_ref.
-    let (client, _statements) = &*pool_ref;
-
-    let st = client
+    let st = conn
         .prepare_typed("SELECT * FROM topics WHERE id=ANY($1)", &[Type::OID_ARRAY])
         .await
         .expect("Failed to prepare");
 
-    let (t, _u) = client
+    let (t, _u) = conn
         .query_raw(&st, [&IDS as &(dyn ToSql + Sync)].iter().map(|s| *s as _))
         .await
         .expect("Failed to query postgres")
@@ -116,14 +113,12 @@ async fn index(
         .await
         .expect("Failed to query postgres");
 
-    drop(pool_ref); // drop the pool_ref when you finish use the pool. so that the connection can be put back to pool asap.
-
     Ok(content::Json(serde_json::to_string(&t).unwrap()))
 }
 
 #[get("/redis")]
 async fn index2(pool: State<'_, Pool<RedisManager>>) -> Result<(), Debug<std::io::Error>> {
-    let mut client = pool.get().await.map_err(MyError::from)?.get_conn().clone();
+    let mut client = pool.get().await.map_err(MyError::from)?.clone();
 
     redis::cmd("PING")
         .query_async::<_, ()>(&mut client)