+++ examples/tide_example/src/main.rs
@@ -48,16 +48,14 @@ async fn main() -> async_std::io::Result<()> {
 
 async fn index(req: Request<State>) -> Result<Response, Error> {
     let state = req.state();
-    let pool_ref = state.postgres_pool.get().await.unwrap();
-    let (client, _statement_map) = &*pool_ref;
+    let conn = state.postgres_pool.get().await.unwrap();
 
-    let _ = client.simple_query("").await.unwrap();
+    let _ = conn.simple_query("").await.unwrap();
 
-    let mut pool_ref = state.redis_pool.get().await.unwrap();
-    let client = &mut *pool_ref;
+    let mut conn = state.redis_pool.get().await.unwrap();
 
     redis::cmd("PING")
-        .query_async::<_, ()>(client)
+        .query_async::<_, ()>(&mut *client)
         .await
         .unwrap();