+++ src/util/mod.rs
@@ -4,6 +4,6 @@ pub(crate) mod backoff;
 #[cfg(feature = "no-send")]
 pub(crate) mod cell_pool;
 pub(crate) mod linked_list;
-pub(crate) mod pool_error;
+pub(crate) mod smart_pointer;
 pub(crate) mod spawn_guard;
 pub(crate) mod timeout;