+++ src/util/spawn_guard.rs
@@ -1,4 +1,4 @@
-use crate::pool_inner::PoolInner;
+use crate::pool::PoolRefBehavior;
 use crate::{Manager, SharedManagedPool};
 
 // use a guard type to monitor the spawn result.
@@ -7,7 +7,7 @@ pub(crate) struct SpawnGuard<'a, M>
 where
     M: Manager,
 {
-    pool_inner: &'a PoolInner<M>,
+    shared_pool: &'a SharedManagedPool<M>,
     fulfilled: bool,
 }
 
@@ -15,15 +15,23 @@ impl<'a, M> SpawnGuard<'a, M>
 where
     M: Manager,
 {
-    pub(crate) fn new(pool_inner: &'a PoolInner<M>) -> Self {
+    pub(crate) fn new(shared_pool: &'a SharedManagedPool<M>) -> Self {
         Self {
-            pool_inner,
+            shared_pool,
             fulfilled: false,
         }
     }
 
-    pub(crate) fn fulfilled(&mut self) {
-        self.fulfilled = true;
+    pub(crate) async fn add<R>(mut self) -> Result<R, M::Error>
+    where
+        R: PoolRefBehavior<'a, M>,
+    {
+        let p = self.shared_pool;
+
+        p.add().await.map(|conn| {
+            self.fulfilled = true;
+            R::from_conn(conn, p)
+        })
     }
 }
 
@@ -33,7 +41,7 @@ where
 {
     fn drop(&mut self) {
         if !self.fulfilled {
-            self.pool_inner.dec_active();
+            self.shared_pool.dec_active();
         }
     }
 }
@@ -51,19 +59,19 @@ impl<M> SpawnGuardOwned<M>
 where
     M: Manager,
 {
-    pub(crate) fn new(shared_pool: SharedManagedPool<M>) -> Self {
+    pub(crate) fn new(shared_pool: &SharedManagedPool<M>) -> Self {
         Self {
-            shared_pool,
+            shared_pool: shared_pool.clone(),
             fulfilled: false,
         }
     }
 
-    pub(crate) fn shared_pool(&self) -> &SharedManagedPool<M> {
-        &self.shared_pool
-    }
-
-    pub(crate) fn fulfilled(&mut self) {
-        self.fulfilled = true;
+    pub(crate) async fn add(mut self) {
+        let p = &self.shared_pool;
+        if let Ok(conn) = p.add().await {
+            p.push_back(conn);
+            self.fulfilled = true;
+        };
     }
 }